+++ src/FluxFlow.Components.Metrics/MetricsComponentRegistrationExtensions.cs
@@ -1,3 +1,4 @@
+using FluxFlow.Components.Metrics.Options;
 using FluxFlow.Engine.Runtime;
 
 namespace FluxFlow.Components.Metrics;
@@ -6,8 +7,17 @@ public static class MetricsComponentRegistrationExtensions
 {
     public static RuntimeNodeFactoryRegistry RegisterMetricsComponents(
         this RuntimeNodeFactoryRegistry registry)
+        => registry.RegisterMetricsComponents(_ => { });
+
+    public static RuntimeNodeFactoryRegistry RegisterMetricsComponents(
+        this RuntimeNodeFactoryRegistry registry,
+        Action<MetricsComponentOptions> configure)
     {
         ArgumentNullException.ThrowIfNull(registry);
-        return registry.Register(new MetricsComponentModule());
+        ArgumentNullException.ThrowIfNull(configure);
+
+        var options = new MetricsComponentOptions();
+        configure(options);
+        return registry.Register(new MetricsComponentModule(options));
     }
 }