+++ src/FluxFlow.Components.Metrics/Timing/IMetricsClock.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.Metrics.Timing;
+
+public interface IMetricsClock
+{
+    DateTimeOffset UtcNow { get; }
+}